Fixes for 0.1.2 (#70)
    – Do not ask the password when selected the same account;
    – Essentia coins default names;
    – Remove scroll to top in settings;
    – Add static settings title;
    – Remove login method;
    – Correct backup and login routes;
    – Add settings scroll background;
    – Scroll freezing on settings screen for crypto;
    – Remove restore from backup xib;
    – Fix settings animation;
    – Fix async reload crash;
    – Pin add wallet navigation bar to top, remove tableview with size component size;
    – Select wallet screen after logged in to another user;
    – Reload asset balance after adding;
    – Fee in current currency.

+++ Essentia/Core/Services/UserService/UserStorageService.swift
@@ -42,7 +42,9 @@ class UserStorageService: UserStorageServiceInterface, AppStateEventHandler {
     
     func remove(user: User) {
         do {
-            try fileSerice.removeFile(at: folderPath, with: user.id)
+            if fileSerice.storageContainFile(at: folderPath, with: user.id) {
+                try fileSerice.removeFile(at: folderPath, with: user.id)
+            }
         } catch {
             (inject() as LoggerServiceInterface).log(error.localizedDescription)
         }
